chore: modularized main.py
